Create client.js
Added the client.js file with the jQuery code to the project's repo. This is used to inject the back-end data into the front-end UI. Added functions for all the 3 forms.  When submitting the data, the response is displayed on the current page instead of redirecting you to a new page.