Update api.js
Completed all the required methods for the forms to work as intended. The books are added to the database. The list of all the books in the "library" collection on the database is retrieved without a hitch. The "commentcount" property displays the correct value based on the "comments" array's length. The comments are posted correctly based on the book _id property. The book delete based on _id works as intended. In case the library is empty will be displayed a "None" status message. The delete all books also works as intended. The project is fully functional. 9 out of 10 user stories are completed and all that remains is to write the functional_tests to fully complete the project.